chore: organize team files
